fix: make server-side hostnames available through `KeymanHosts` (#53)
For local environment, we need to use host.docker.internal as the
hostname, with corresponding ports, in order for the sites in
their containers be able to communicate with each other. This adds
`$SERVER_keyman_com` variables etc to the `KeymanHosts` class, in
preparation for fixing this across each of the sites.

This is also groundwork for allowing the various containers in the
production and staging environments to communicate with each other
directly, rather than going via cloudflare, which should improve
performance and reduce traffic on the cluster. This change will be
implemented in a follow-up PR.