feat: simpler _common file specification
Make it easier to add files to _common by removing the need to change
the bootstrap script, instead putting the list of files to be downloaded
into the file /.bootstrap-registry. The corresponding /build.sh will
rebuild /.bootstrap-registry, and should be run any time a file is added
or removed from /_common.