feat: simpler _common file specification (#55)
* feat: simpler _common file specification

Make it easier to add files to _common by removing the need to change
the bootstrap script, instead putting the list of files to be downloaded
into the file /.bootstrap-registry. The corresponding /build.sh will
rebuild /.bootstrap-registry, and should be run any time a file is added
or removed from /_common.

* feat: cdn support for assets

Adds sha1sum of files in the assets folder and makes a copy so that
these files can be used in place of the bare filenames, avoiding cache
issues. Future possibilities of sharing these files.

* feat: add Assets class to lookup CDN asset names

* feat: support cdn files in bootstrap

---------

Co-authored-by: Darcy Wong <darcy_wong@sil.org>